+++ crates/repl/src/runtime_panel.rs
@@ -34,13 +34,6 @@ actions!(
 );
 actions!(repl_panel, [ToggleFocus]);
 
-pub enum SessionSupport {
-    ActiveSession(View<Session>),
-    Inactive(Box<KernelSpecification>),
-    RequiresSetup(Arc<str>),
-    Unsupported,
-}
-
 pub fn init(cx: &mut AppContext) {
     cx.observe_new_views(
         |workspace: &mut Workspace, _cx: &mut ViewContext<Workspace>| {
@@ -84,7 +77,7 @@ pub fn init(cx: &mut AppContext) {
 
                     let weak_editor = cx.view().downgrade();
                     cx.defer(|panel, cx| {
-                        RuntimePanel::run(weak_editor, cx).log_err();
+                        crate::editor::run(weak_editor, cx).log_err();
                     });
                 },
             ))
@@ -99,7 +92,7 @@ pub fn init(cx: &mut AppContext) {
 
                     let weak_editor = cx.view().downgrade();
                     cx.defer(|panel, cx| {
-                        RuntimePanel::clear_outputs(weak_editor, cx);
+                        crate::editor::clear_outputs(weak_editor, cx);
                     });
                 },
             ))
@@ -114,7 +107,7 @@ pub fn init(cx: &mut AppContext) {
 
                     let weak_editor = cx.view().downgrade();
                     cx.defer(|panel, cx| {
-                        RuntimePanel::interrupt(weak_editor, cx);
+                        crate::editor::interrupt(weak_editor, cx);
                     });
                 },
             ))
@@ -129,7 +122,7 @@ pub fn init(cx: &mut AppContext) {
 
                     let weak_editor = cx.view().downgrade();
                     cx.defer(|panel, cx| {
-                        RuntimePanel::shutdown(weak_editor, cx);
+                        crate::editor::shutdown(weak_editor, cx);
                     });
                 },
             ))
@@ -203,168 +196,6 @@ impl RuntimePanel {
 
     fn focus_out(&mut self, _event: FocusOutEvent, cx: &mut ViewContext<Self>) {
         cx.notify();
-    }
-
-    fn snippet(
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Editor>,
-    ) -> Option<(String, Arc<Language>, Range<Anchor>)> {
-        let editor = editor.upgrade()?;
-        let editor = editor.read(cx);
-
-        let buffer = editor.buffer().read(cx).snapshot(cx);
-
-        let selection = editor.selections.newest::<usize>(cx);
-        let multi_buffer_snapshot = editor.buffer().read(cx).snapshot(cx);
-
-        let range = if selection.is_empty() {
-            let cursor = selection.head();
-
-            let cursor_row = multi_buffer_snapshot.offset_to_point(cursor).row;
-            let start_offset = multi_buffer_snapshot.point_to_offset(Point::new(cursor_row, 0));
-
-            let end_point = Point::new(
-                cursor_row,
-                multi_buffer_snapshot.line_len(MultiBufferRow(cursor_row)),
-            );
-            let end_offset = start_offset.saturating_add(end_point.column as usize);
-
-            // Create a range from the start to the end of the line
-            start_offset..end_offset
-        } else {
-            selection.range()
-        };
-
-        let anchor_range = range.to_anchors(&multi_buffer_snapshot);
-
-        let selected_text = buffer
-            .text_for_range(anchor_range.clone())
-            .collect::<String>();
-
-        let start_language = buffer.language_at(anchor_range.start)?;
-        let end_language = buffer.language_at(anchor_range.end)?;
-        if start_language != end_language {
-            return None;
-        }
-
-        Some((selected_text, start_language.clone(), anchor_range))
-    }
-
-    fn language(editor: WeakView<Editor>, cx: &mut AppContext) -> Option<Arc<Language>> {
-        let editor = editor.upgrade()?;
-        let selection = editor.read(cx).selections.newest::<usize>(cx);
-        let buffer = editor.read(cx).buffer().read(cx).snapshot(cx);
-        buffer.language_at(selection.head()).cloned()
-    }
-
-    pub fn run(editor: WeakView<Editor>, cx: &mut ViewContext<Editor>) -> Result<()> {
-        let store = ReplStore::global(cx);
-        if !store.read(cx).is_enabled() {
-            return Ok(());
-        }
-
-        let (selected_text, language, anchor_range) = match Self::snippet(editor.clone(), cx) {
-            Some(snippet) => snippet,
-            None => return Ok(()),
-        };
-
-        let entity_id = editor.entity_id();
-
-        let kernel_specification = store.update(cx, |store, cx| {
-            store
-                .kernelspec(&language, cx)
-                .with_context(|| format!("No kernel found for language: {}", language.name()))
-        })?;
-
-        let fs = store.read(cx).fs().clone();
-        let session = if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
-            session
-        } else {
-            let session = cx.new_view(|cx| Session::new(editor, fs, kernel_specification, cx));
-            cx.notify();
-
-            let subscription = cx.subscribe(&session, {
-                let store = store.clone();
-                move |_this, _session, event, cx| match event {
-                    SessionEvent::Shutdown(shutdown_event) => {
-                        store.update(cx, |store, _cx| {
-                            store.remove_session(shutdown_event.entity_id());
-                        });
-                    }
-                }
-            });
-
-            subscription.detach();
-
-            store.update(cx, |store, _cx| {
-                store.insert_session(entity_id, session.clone());
-            });
-
-            session
-        };
-
-        session.update(cx, |session, cx| {
-            session.execute(&selected_text, anchor_range, cx);
-        });
-
-        anyhow::Ok(())
-    }
-
-    pub fn session(editor: WeakView<Editor>, cx: &mut AppContext) -> SessionSupport {
-        let store = ReplStore::global(cx);
-        let entity_id = editor.entity_id();
-
-        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
-            return SessionSupport::ActiveSession(session);
-        };
-
-        let language = Self::language(editor, cx);
-        let language = match language {
-            Some(language) => language,
-            None => return SessionSupport::Unsupported,
-        };
-        let kernelspec = store.update(cx, |store, cx| store.kernelspec(&language, cx));
-
-        match kernelspec {
-            Some(kernelspec) => SessionSupport::Inactive(Box::new(kernelspec)),
-            None => match language.name().as_ref() {
-                "TypeScript" | "Python" => SessionSupport::RequiresSetup(language.name()),
-                _ => SessionSupport::Unsupported,
-            },
-        }
-    }
-
-    pub fn clear_outputs(editor: WeakView<Editor>, cx: &mut ViewContext<Editor>) {
-        let store = ReplStore::global(cx);
-        let entity_id = editor.entity_id();
-        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
-            session.update(cx, |session, cx| {
-                session.clear_outputs(cx);
-            });
-            cx.notify();
-        }
-    }
-
-    pub fn interrupt(editor: WeakView<Editor>, cx: &mut ViewContext<Editor>) {
-        let store = ReplStore::global(cx);
-        let entity_id = editor.entity_id();
-        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
-            session.update(cx, |session, cx| {
-                session.interrupt(cx);
-            });
-            cx.notify();
-        }
-    }
-
-    pub fn shutdown(editor: WeakView<Editor>, cx: &mut ViewContext<Editor>) {
-        let store = ReplStore::global(cx);
-        let entity_id = editor.entity_id();
-        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
-            session.update(cx, |session, cx| {
-                session.shutdown(cx);
-            });
-            cx.notify();
-        }
     }
 }
